+++ hahomematic/model/device.py
@@ -45,10 +45,11 @@ from hahomematic.const import (
     ProductGroup,
     RxMode,
 )
+from hahomematic.decorators import service
 from hahomematic.exceptions import BaseHomematicException, HaHomematicException
 from hahomematic.model.custom import data_point as hmce, definition as hmed
 from hahomematic.model.data_point import BaseParameterDataPoint, CallbackDataPoint
-from hahomematic.model.decorators import info_property, service, state_property
+from hahomematic.model.decorators import info_property, state_property
 from hahomematic.model.event import GenericEvent
 from hahomematic.model.generic import GenericDataPoint
 from hahomematic.model.support import (
@@ -514,6 +515,7 @@ class Device(PayloadMixin):
                 f"EXPORT_DEVICE_DEFINITION failed: {reduce_args(args=ex.args)}"
             ) from ex
 
+    @service()
     def refresh_firmware_data(self) -> None:
         """Refresh firmware data of the device."""
         old_available_firmware = self.available_firmware
@@ -534,6 +536,7 @@ class Device(PayloadMixin):
             for callback_handler in self._firmware_update_callbacks:
                 callback_handler()
 
+    @service()
     async def update_firmware(self, refresh_after_update_intervals: tuple[int, ...]) -> bool:
         """Update the firmware of the homematic device."""
         update_result = await self._client.update_device_firmware(device_address=self._address)
@@ -548,6 +551,7 @@ class Device(PayloadMixin):
 
         return update_result  # type: ignore[no-any-return]
 
+    @service()
     async def load_value_cache(self) -> None:
         """Init the parameter cache."""
         if len(self.generic_data_points) > 0:
@@ -559,6 +563,7 @@ class Device(PayloadMixin):
             self._address,
         )
 
+    @service()
     async def reload_paramset_descriptions(self) -> None:
         """Reload paramset for device."""
         for (
@@ -744,6 +749,7 @@ class Channel(PayloadMixin):
         """Return the unique_id of the channel."""
         return self._unique_id
 
+    @service()
     async def create_central_link(self) -> None:
         """Create a central link to support press events."""
         if self._has_key_press_events and not await self._has_central_link():
@@ -751,6 +757,7 @@ class Channel(PayloadMixin):
                 address=self._address, value_id=REPORT_VALUE_USAGE_VALUE_ID, ref_counter=1
             )
 
+    @service()
     async def remove_central_link(self) -> None:
         """Remove a central link."""
         if (
@@ -762,6 +769,7 @@ class Channel(PayloadMixin):
                 address=self._address, value_id=REPORT_VALUE_USAGE_VALUE_ID, ref_counter=0
             )
 
+    @service()
     async def cleanup_central_link_metadata(self) -> None:
         """Cleanup the metadata for central links."""
         if metadata := await self._device.client.get_metadata(
@@ -1141,6 +1149,7 @@ class _DefinitionExporter:
         self._device_address: Final = device.address
         self._random_id: Final[str] = f"VCU{int(random.randint(1000000, 9999999))}"
 
+    @service()
     async def export_data(self) -> None:
         """Export data."""
         device_descriptions: Mapping[str, DeviceDescription] = (
